+++ .vscode/c_cpp_properties.json
@@ -3,7 +3,8 @@
         {
             "name": "Linux",
             "includePath": [
-                "${workspaceFolder}/**"
+                "${workspaceFolder}/**",
+                "${workspaceFolder}/pkg/message/**"
             ],
             "defines": [],
             "compilerPath": "/usr/bin/gcc",
