--- dormitory.cpp
@@ -1,1 +1,0 @@
-#include "dormitory.hpp"
