--- mqtt.hpp
@@ -1,4 +1,0 @@
-#ifndef __MQTT_HPP__
-#define __MQTT_HPP__
-
-#endif // __MQTT_HPP__
