+++ pkg/message/mqtt.hpp
@@ -1,0 +1,4 @@
+#ifndef __MQTT_HPP__
+#define __MQTT_HPP__
+
+#endif // __MQTT_HPP__
